Commit Message: ``` Add watermarking functionality to script
This commit adds a zsh script that utilizes ImageMagick to apply a watermark to a specified directory of images. The script includes configuration options for the watermark image, input and output directories, position, and padding. It also checks for the existence of the output directory and creates it if necessary. The script then loops through all images in the input directory, applies the watermark using the convert command from ImageMagick, and saves the watermarked images in the output directory. Upon completion, it displays a message indicating the watermarking is finished and provides the location of the output directory.
```