Commit Message: "Add Batch Image Watermarking Script"
Description:
Created a zsh script for batch image watermarking, utilizing the ImageMagick tool. The script checks for the existence of the watermark image and creates an output directory for watermarked images if it does not exist. It then applies the watermark to all JPEG and PNG images in the current directory, saving the watermarked images in the designated output directory.