Commit Message: ``` feat: Add script to create bulk users from userlist.txt
Adds a zsh script that reads a list of usernames from userlist.txt and creates new users with a default password. The script logs the user creation process to user_creation.log. If a user already exists, it skips the creation process for that user. Upon completion, the script logs a message indicating successful completion of the bulk user creation process.
```