"Add script to synchronize time on network devices
This commit introduces a zsh script that defines an array of network device IPs or hostnames and a function to synchronize time on a network device using ssh. It iterates over the list of network devices and syncs the time for each device by executing the `ntpdate` command on the remote device using ssh. "