Commit message for the script: ``` feat: Create script to create user accounts from a provided list
Add a script that reads a list of usernames from a file, checks if the users already exist, and creates new user accounts with default permissions if they do not exist. The script uses `useradd` to create the users with a default password and sets the home directory permissions to 700 (rwx only for the user).
```