"Add Python script to generate FAQ response from a template
The script defines sample FAQs data and a Jinja template for the FAQ response. It includes a function to generate the FAQ response based on the template and the provided data. The resulting FAQ response is then written to an HTML file. This script can be used to easily generate and customize FAQ responses based on different datasets and templates."